Added more tests and a simple example project

+++ QuickPay/Models/Payments/Basket.cs
@@ -5,7 +5,7 @@
         public int qty { get; set; }
         public string item_no { get; set; }
         public string item_name { get; set; }
-        public int item_price { get; set; }
+        public double item_price { get; set; }
         public double vat_rate { get; set; }
     }
 }
